Completed the assignment, fully functional
