+++ Bank.cs
@@ -1,0 +1,9 @@
+namespace LINQed
+{
+    public class Bank
+    {
+        public string Symbol { get; set; }
+        public string Name { get; set; }
+    }
+
+}
