+++ Program.cs
@@ -6,6 +6,27 @@ namespace LINQed
 {
     class Program
     {
+
+        static List<Customer> customers = new List<Customer>() {
+                    new Customer(){ Name="Bob Lesman", Balance=80345.66, Bank="FTB"},
+                    new Customer(){ Name="Joe Landy", Balance=9284756.21, Bank="WF"},
+                    new Customer(){ Name="Meg Ford", Balance=487233.01, Bank="BOA"},
+                    new Customer(){ Name="Peg Vale", Balance=7001449.92, Bank="BOA"},
+                    new Customer(){ Name="Mike Johnson", Balance=790872.12, Bank="WF"},
+                    new Customer(){ Name="Les Paul", Balance=8374892.54, Bank="WF"},
+                    new Customer(){ Name="Sid Crosby", Balance=957436.39, Bank="FTB"},
+                    new Customer(){ Name="Sarah Ng", Balance=56562389.85, Bank="FTB"},
+                    new Customer(){ Name="Tina Fey", Balance=1000000.00, Bank="CITI"},
+                    new Customer(){ Name="Sid Brown", Balance=49582.68, Bank="CITI"}
+                };
+
+        static List<Bank> banks = new List<Bank>() {
+                    new Bank(){ Name="First Tennessee", Symbol="FTB"},
+                    new Bank(){ Name="Wells Fargo", Symbol="WF"},
+                    new Bank(){ Name="Bank of America", Symbol="BOA"},
+                    new Bank(){ Name="Citibank", Symbol="CITI"},
+                };
+
         static void Main(string[] args)
         {
             Console.WriteLine("Hello World!");
@@ -91,98 +112,63 @@ namespace LINQed
                 FTB 1
                 CITI 1
             */
+            // Build a collection of customers who are millionaires
+            
             customers
                 .Where(Customer => Customer.Balance > 999999.99)
                 .GroupBy(Customer => Customer.Bank).ToList()
                 .ForEach(x => Console.WriteLine($"{x.Key} has {x.Count()} millionaires!"));
 
+
+                /* As a light introduction to working with relational databases, this example works with two collections of 
+                data - `banks` and `customers` - that are related through the `Bank` attribute on the customer. In that 
+                attribute, we store the abbreviation for a bank. However, we want to get the full name of the bank when we 
+                produce our output.
+
+                This is called joining the collections together.
+
+                This exercise is also an introduction to producing anonymous objects as the result of the LINQ statement.
+
+                Read the [Group Join](https://code.msdn.microsoft.com/LINQ-Join-Operators-dabef4e9#groupjoin) example to 
+                get started. */
+
+                /*
+                    TASK:
+                    As in the previous exercise, you're going to output the millionaires,
+                    but you will also display the full name of the bank. You also need
+                    to sort the millionaires' names, ascending by their LAST name.
+
+                    Example output:
+                        Tina Fey at Citibank
+                        Joe Landy at Wells Fargo
+                        Sarah Ng at First Tennessee
+                        Les Paul at Wells Fargo
+                        Peg Vale at Bank of America
+                */
+
+                // Define a bank
+
+                // Create some banks and store in a List
+
+                // var millionaireReport = ...
+
+                // foreach (var customer in millionaireReport)
+                //     {
+                //         Console.WriteLine($"{customer.Name} at {customer.Bank}");
+                //     }
+
             banks
                 .Join(customers,
                 b => b.Symbol,
                 c => c.Bank,
-                (customer, bank) => new {customer.Name, bank.Name}
+                (bank, customer) => new {CustomerName = customer.Name, CustomberBalance = customer.Balance, BankName = bank.Name}
                 )
-                .Where(c => c.Balance > 999999.99)
-                .ForEach(CustomerAtBank => Console.WriteLine($"{CustomerAtBank.customer.name} at {CustomerAtBank.bank.name} ");
-            // customers
-            //     .Where(Customer => Customer.Balance > 999999.99)
-            //     .Join(Customer => Customer.Bank == Banks.Symbol).ToList()
-            //     .ForEach(y => Console.WriteLine($"{y.customer.name} at {y.bank.name}"));
+                .Where(CustomerInfo => CustomerInfo.CustomberBalance > 999999.99)
+                .OrderBy(Customer => Customer.CustomerName.Split(' ')[1]).ToList()
+                .ForEach(CustomerAtBank => Console.WriteLine($"{CustomerAtBank.CustomerName} at {CustomerAtBank.BankName}"));
         }
 
 
-        // Build a collection of customers who are millionaires
-        public class Customer
-        {
-            public string Name { get; set; }
-            public double Balance { get; set; }
-            public string Bank { get; set; }
-        }
 
-        static List<Customer> customers = new List<Customer>() {
-                    new Customer(){ Name="Bob Lesman", Balance=80345.66, Bank="FTB"},
-                    new Customer(){ Name="Joe Landy", Balance=9284756.21, Bank="WF"},
-                    new Customer(){ Name="Meg Ford", Balance=487233.01, Bank="BOA"},
-                    new Customer(){ Name="Peg Vale", Balance=7001449.92, Bank="BOA"},
-                    new Customer(){ Name="Mike Johnson", Balance=790872.12, Bank="WF"},
-                    new Customer(){ Name="Les Paul", Balance=8374892.54, Bank="WF"},
-                    new Customer(){ Name="Sid Crosby", Balance=957436.39, Bank="FTB"},
-                    new Customer(){ Name="Sarah Ng", Balance=56562389.85, Bank="FTB"},
-                    new Customer(){ Name="Tina Fey", Balance=1000000.00, Bank="CITI"},
-                    new Customer(){ Name="Sid Brown", Balance=49582.68, Bank="CITI"}
-                };
-
-        List<Bank> banks = new List<Bank>() {
-            new Bank(){ Name="First Tennessee", Symbol="FTB"},
-            new Bank(){ Name="Wells Fargo", Symbol="WF"},
-            new Bank(){ Name="Bank of America", Symbol="BOA"},
-            new Bank(){ Name="Citibank", Symbol="CITI"},
-        };
-
-
-        /* As a light introduction to working with relational databases, this example works with two collections of 
-        data - `banks` and `customers` - that are related through the `Bank` attribute on the customer. In that 
-        attribute, we store the abbreviation for a bank. However, we want to get the full name of the bank when we 
-        produce our output.
-
-        This is called joining the collections together.
-
-        This exercise is also an introduction to producing anonymous objects as the result of the LINQ statement.
-
-        Read the [Group Join](https://code.msdn.microsoft.com/LINQ-Join-Operators-dabef4e9#groupjoin) example to 
-        get started. */
-
-        /*
-            TASK:
-            As in the previous exercise, you're going to output the millionaires,
-            but you will also display the full name of the bank. You also need
-            to sort the millionaires' names, ascending by their LAST name.
-
-            Example output:
-                Tina Fey at Citibank
-                Joe Landy at Wells Fargo
-                Sarah Ng at First Tennessee
-                Les Paul at Wells Fargo
-                Peg Vale at Bank of America
-        */
-
-        // Define a bank
-
-        // Create some banks and store in a List
-
-        // var millionaireReport = ...
-
-        // foreach (var customer in millionaireReport)
-        //     {
-        //         Console.WriteLine($"{customer.Name} at {customer.Bank}");
-        //     }
     }
-
-
-    public class Bank
-    {
-        public string Symbol { get; set; }
-        public string Name { get; set; }
-    }
-
 }
